requirements.txt
haystack-ai
pinecone-haystack
amazon-bedrock-haystack
python-dotenv
beautifulsoup4
boto3
uvicorn
fastapi
pydantic
requests
streamlit